move active storage downloader to concern
